+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Fixed
+- Make stored node proposals immutable
 
 ## [v2.5.4] - 2020-04-22
 ### Added